Add missing includes.
Fixes build without PCH.
