+++ src/game/AI/PlayerAI.cpp
@@ -23,6 +23,7 @@
 #include "SpellMgr.h"
 #include "MotionMaster.h"
 #include "Spell.h"
+#include "Map.h"
 
 void PlayerAI::Remove()
 {
