+++ src/game/Formulas.h
@@ -25,6 +25,7 @@
 #include "World.h"
 #include "Creature.h"
 #include "Player.h"
+#include "Map.h"
 
 namespace MaNGOS
 {
