+++ src/game/Guild/GuildMgr.h
@@ -22,6 +22,7 @@
 #include "Common.h"
 #include "Policies/Singleton.h"
 #include "World.h"
+#include <shared_mutex>
 
 class Guild;
 class ObjectGuid;
