+++ src/game/PacketBroadcast/MovementBroadcaster.h
@@ -3,7 +3,7 @@
 
 #include <atomic>
 #include <chrono>
-#include <mutex>
+#include <shared_mutex>
 #include <thread>
 #include <vector>
 #include <cstddef>
